fixed: blog 컴포넌트 util 분리
